structures for definition code (Tests and Challenges)
JIRA: AUTO-13

Created structures for configuring various test definition scripts
and challenge definition scripts in corresponding classes (something like
libraries of functions, with an index ID to pick the applicable function).
This will be useful to leverage Python clients to OpenStack and
Kubernetes, or even for HTTP-level APIs for ONAP. It may be more flexible
than the list of CLI/API commands initially planned (however, the list structures
are still there, and might be used by the scripts).

Additionally, there are some test snippets for Openstack Python SDK (work in progress).
Also, there is a clouds.yaml file, for this interface.

Change-Id: Ida1ff118fcbfe8715679492feb8fe0df2da62b45
Signed-off-by: Gerard Damm <gerard.damm@wipro.com>