Add Gerard fix meeting time.
Any comitters that are listed in the INFO file
and not in the INFO.yaml never accepted their invitations

They should perhaps be removed from the committer list.
and email asking them to stepdown forwarded to helpdesk is the correct
procedure to remove inactive comitters.

Or if they are interested I can resent the invite.

Change-Id: Id5550db13a8f64a02934bf1e106d228a76fdc15e
Signed-off-by: agardner <agardner@linuxfoundation.org>