automated configuration of an OpenStack instance for ONAP
Initial code (based on OpenStack SDK Python client) for configuring
an OpenStack instance (e.g. installed by Fuel/MCP) and prepare it for
ONAP (especially, to host VM-based VNFs).

Currently works: creation of networks, routers, router interfaces,
security groups (accessing VMs also works with compute proxy, but is
not used for this program).

Current issues: identity proxy (so can't create projects and users),
adding gateway to router, deletions

This script could be launched automatically right after installing
OpenStack, and before installing ONAP (so as to grab OpenStack info
to be fed into ONAP configuration files). Also, VNF YAML/ENV files
can be populated once the OpenStack instance is ready for ONAP.

Note: plenty of test and debug lines, which would be removed.
This changeset is just to capture a snapshot of this effort in the
Auto repository, and to enable reviews and feedback.

Change-Id: If3b69b5a9a45905455db3a4eae0c6b336ce7447b
Signed-off-by: Gerard Damm <gerard.damm@wipro.com>