Update patch set 1
Patch Set 1:

(5 comments)

Hi Christina,

thanks a lot for your review. It is to note, that this script is a work in progress. It was kept as attachment in jira, which was not very flexible. The idea is to keep the track of its changes in the gerrit, so other interested team members can easily push it into their environment and test it, etc.

Patch-set: 1
Reviewer: Gerrit User 1306 <1306@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0